Refactor supabase env variables
